abi: Enable support for 64-bit ABI
This change updates ABI in Hubris to support 64-bit. This involves
two pieces:
1. Host
2. Runtime

For runtime, it is straightforward since `usize` can be used to
determine how many bytes it takes to reference any memory location
and the registers.

For host, `usize` does not provide this information. Thus, we have to
instead rely on inputs from user to determine what ABI should be used
for building any host tools e.g. xtask.

Thus, sys/abi is updated to define AbiSize as one of the following:
1. usize --> Required for runtime by kernel and user tasks
2. u32   --> Required for host tools that use 32-bit ABI
3. u64   --> Required for host tools that use 64-bit ABI

In order to enable the above definitions, following two feature flags
are added that can be selected by the user when building host tools:
`host_use_abi32` and `host_use_abi64`. This change updates sys/abi and
all dependent host tools (xtask), kern and user tasks (including tests)
to update the data types used in ABI calls.

+++ sys/kern/src/arch/arm_m.rs
@@ -159,55 +159,55 @@ pub struct SavedState {
 /// Map the volatile registers to (architecture-independent) syscall argument
 /// and return slots.
 impl task::ArchState for SavedState {
-    fn stack_pointer(&self) -> u32 {
-        self.psp
+    fn stack_pointer(&self) -> usize {
+        self.psp as usize
     }
 
     /// Reads syscall argument register 0.
-    fn arg0(&self) -> u32 {
-        self.r4
+    fn arg0(&self) -> usize {
+        self.r4 as usize
     }
-    fn arg1(&self) -> u32 {
-        self.r5
+    fn arg1(&self) -> usize {
+        self.r5 as usize
     }
-    fn arg2(&self) -> u32 {
-        self.r6
+    fn arg2(&self) -> usize {
+        self.r6 as usize
     }
-    fn arg3(&self) -> u32 {
-        self.r7
+    fn arg3(&self) -> usize {
+        self.r7 as usize
     }
-    fn arg4(&self) -> u32 {
-        self.r8
+    fn arg4(&self) -> usize {
+        self.r8 as usize
     }
-    fn arg5(&self) -> u32 {
-        self.r9
+    fn arg5(&self) -> usize {
+        self.r9 as usize
     }
-    fn arg6(&self) -> u32 {
-        self.r10
+    fn arg6(&self) -> usize {
+        self.r10 as usize
     }
 
-    fn syscall_descriptor(&self) -> u32 {
-        self.r11
+    fn syscall_descriptor(&self) -> usize {
+        self.r11 as usize
     }
 
     /// Writes syscall return argument 0.
-    fn ret0(&mut self, x: u32) {
-        self.r4 = x
+    fn ret0(&mut self, x: usize) {
+        self.r4 = x as u32
     }
-    fn ret1(&mut self, x: u32) {
-        self.r5 = x
+    fn ret1(&mut self, x: usize) {
+        self.r5 = x as u32
     }
-    fn ret2(&mut self, x: u32) {
-        self.r6 = x
+    fn ret2(&mut self, x: usize) {
+        self.r6 = x as u32
     }
-    fn ret3(&mut self, x: u32) {
-        self.r7 = x
+    fn ret3(&mut self, x: usize) {
+        self.r7 = x as u32
     }
-    fn ret4(&mut self, x: u32) {
-        self.r8 = x
+    fn ret4(&mut self, x: usize) {
+        self.r8 = x as u32
     }
-    fn ret5(&mut self, x: u32) {
-        self.r9 = x
+    fn ret5(&mut self, x: usize) {
+        self.r9 = x as u32
     }
 }
 
@@ -294,7 +294,7 @@ pub fn reinitialize(task: &mut task::Task) {
             continue;
         }
 
-        let mut uslice: USlice<u32> = USlice::from_raw(
+        let mut uslice: USlice<usize> = USlice::from_raw(
             region.base as usize,
             (initial_stack as usize - frame_size - region.base as usize) >> 2,
         )
@@ -312,7 +312,7 @@ pub fn reinitialize(task: &mut task::Task) {
     // Conservatively/defensively zero the entire frame.
     *frame = ExtendedExceptionFrame::default();
     // Now fill in the bits we actually care about.
-    frame.base.pc = descriptor.entry_point | 1; // for thumb
+    frame.base.pc = descriptor.entry_point as u32 | 1; // for thumb
     frame.base.xpsr = INITIAL_PSR;
     frame.base.lr = 0xFFFF_FFFF; // trap on return from main
     #[cfg(any(armv7m, armv8m))]
@@ -342,7 +342,7 @@ pub fn apply_memory_protection(task: &task::Task) {
     for (i, region) in task.region_table().iter().enumerate() {
         let rbar = (i as u32)  // region number
             | (1 << 4)  // honor the region number
-            | region.base;
+            | region.base as u32;
         let ratts = region.attributes;
         let xn = !ratts.contains(abi::RegionAttributes::EXECUTE);
         // These AP encodings are chosen such that we never deny *privileged*
@@ -475,7 +475,7 @@ pub fn apply_memory_protection(task: &task::Task) {
         };
 
         // RLAR = our upper bound
-        let rlar = (region.base + region.size)
+        let rlar = (region.base + region.size) as u32
                 | (i as u32) << 1 // AttrIndx
                 | (1 << 0); // enable
 
@@ -483,7 +483,7 @@ pub fn apply_memory_protection(task: &task::Task) {
         let rbar = (xn as u32)
             | ap << 1
             | (sh as u32) << 3  // sharability
-            | region.base;
+            | region.base as u32;
 
         unsafe {
             mpu.rnr.write(rnr);
@@ -1504,14 +1504,19 @@ unsafe extern "C" fn handle_fault(
                 // fact that the user's stack pointer is so trashed that we
                 // can't store through it.  (In particular, we seem to have no
                 // way at getting at our faulted PC.)
-                (FaultInfo::StackOverflow { address: psp }, true)
+                (
+                    FaultInfo::StackOverflow {
+                        address: psp as usize,
+                    },
+                    true,
+                )
             } else if cfsr.contains(Cfsr::IACCVIOL) {
                 (FaultInfo::IllegalText, false)
             } else {
                 (
                     FaultInfo::MemoryAccess {
                         address: if cfsr.contains(Cfsr::MMARVALID) {
-                            Some(scb.mmfar.read())
+                            Some(scb.mmfar.read() as usize)
                         } else {
                             None
                         },
@@ -1525,7 +1530,7 @@ unsafe extern "C" fn handle_fault(
         FaultType::BusFault => (
             FaultInfo::BusError {
                 address: if cfsr.contains(Cfsr::BFARVALID) {
-                    Some(scb.bfar.read())
+                    Some(scb.bfar.read() as usize)
                 } else {
                     None
                 },
