+++ sys/userlib/src/lib.rs
@@ -64,8 +64,8 @@ impl<'a> Lease<'a> {
         Self {
             _kern_rep: abi::ULease {
                 attributes: abi::LeaseAttributes::READ,
-                base_address: x.as_ptr() as u32,
-                length: x.len() as u32,
+                base_address: x.as_ptr() as usize,
+                length: x.len() as usize,
             },
             _marker: PhantomData,
         }
@@ -75,8 +75,8 @@ impl<'a> Lease<'a> {
         Self {
             _kern_rep: abi::ULease {
                 attributes: LeaseAttributes::READ | LeaseAttributes::WRITE,
-                base_address: x.as_ptr() as u32,
-                length: x.len() as u32,
+                base_address: x.as_ptr() as usize,
+                length: x.len() as usize,
             },
             _marker: PhantomData,
         }
@@ -86,8 +86,8 @@ impl<'a> Lease<'a> {
         Self {
             _kern_rep: abi::ULease {
                 attributes: LeaseAttributes::WRITE,
-                base_address: x.as_ptr() as u32,
-                length: x.len() as u32,
+                base_address: x.as_ptr() as usize,
+                length: x.len() as usize,
             },
             _marker: PhantomData,
         }
